% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_signal_canvas.R
\name{add_signal_canvas}
\alias{add_signal_canvas}
\title{Add signal canvases to EMU database}
\usage{
add_signal_canvas(db_handle, add)
}
\arguments{
\item{db_handle}{The handle of an EMU database which is already loaded in R.}

\item{add}{One or more strings containing the names of canvases to add to
the EMU database.}
}
\description{
Wrapper to add one or more signal canvases to the default view of an EMU
database.
}
\examples{
\dontrun{
datapath <- system.file('extdata/db', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
emuR::get_signalCanvasesOrder(raw, 'default')
add_signal_canvas(db_handle=raw, add='eggF0')
emuR::get_signalCanvasesOrder(raw, 'default')
#To have a look at the new order in EMU-SDMS, run
emuR::serve(raw)
}
}
\seealso{
This function wraps the functions \link[emuR:SetGetSignalCanvasesOrder]{emuR::get_signalCanvasesOrder}
and \link[emuR:SetGetSignalCanvasesOrder]{emuR::set_signalCanvasesOrder}.
}
