% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_run_praatsauce.R
\name{bulk_run_praatsauce}
\alias{bulk_run_praatsauce}
\title{Run PraatSauce in bulk with different arguments for different speakers}
\usage{
bulk_run_praatsauce(directory, dyn_args, ...)
}
\arguments{
\item{directory}{Root directory containing subdirectories for individual
speakers to be measured. Must be the full name, there's no
path extension.}

\item{dyn_args}{A named list or data frame containing dynamic arguments. If a
list is provided, the first element should be a vector called \code{speakers}, and
other elements should be vectors with arguments passed to \code{\link[=run_praatsauce]{run_praatsauce()}}.
If a data frame is provided, the first column should be called \code{speakers},
and the other columns should have named that can be passed to
\code{\link[=run_praatsauce]{run_praatsauce()}}.}

\item{...}{Other arguments passed to \code{\link[=run_praatsauce]{run_praatsauce()}}, which are the
same for all speakers.}
}
\value{
A data frame with columns giving information about the
sound file imported from the name of the sound file, information about
the time at which measures were estimated, and the estimated measures
themselves.
}
\description{
Shell wrapper to run PraatSauce in bulk with different arguments for
different speakers, fx with varying pitch floor and ceiling or varying
formant references.
}
\examples{
\dontrun{
da <- list(speakers=c('a', 'b', 'c'), f0min=c(120, 150, 80),
f0max=c(300, 400, 250))
bulk_run_praatsauce(directory='my_directory', dyn_args=da)
}
}
\seealso{
As mentioned above, this function is a wrapper for a suite of
Praat scripts. More information about the scripts can be found
on GitHub: \url{https://github.com/kirbyj/praatsauce}. Information about the
VoiceSauce scripts which PraatSauce is based on can be found here:
\url{http://www.phonetics.ucla.edu/voicesauce/}.

The function \code{\link[=run_ps_dynamic_minmax]{run_ps_dynamic_minmax()}} will dynamically estimate suitable
pitch floors and ceilings for a number of speakers.

The function \code{\link[=praatsauce2ssff]{praatsauce2ssff()}} will convert measures from PraatSauce to
SSFF files and add them to EMU database.
}
