% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_spec.R
\name{clear_spec}
\alias{clear_spec}
\title{No longer display SSFF track on top of spectrogram when viewing EMU database}
\usage{
clear_spec(db, perspective = "default")
}
\arguments{
\item{db}{The handle of an EMU database which is loaded into R.}

\item{perspective}{A string giving the name of the perspective for which the
configuration should be changed. Default is \code{default}.}
}
\description{
Change the configuration of an EMU database perspective to no longer
display signal from an SSFF track on top of the spectrogram.
}
\examples{
\dontrun{
datapath <- system.file('extdata/db', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
emuR::list_ssffTrackDefinitions(raw)

# Add signal
signal_on_spec(db, track='praatF0')
emuR::serve(raw)

# Remove signal
clear_spec(db)
emuR::serve(raw)
}
}
\seealso{
\code{signal_on_spec()} is a function for adding SSFF tracks to a
spectrogram when viewing an EMU database
}
