% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_minmax.R
\name{dynamic_minmax}
\alias{dynamic_minmax}
\title{Helper function for running PraatSauce with dynamic pitch floor and ceiling}
\usage{
dynamic_minmax(fm = TRUE, sm = TRUE, ...)
}
\arguments{
\item{fm}{Boolean; should formants be measured in the second pass?}

\item{sm}{Boolean; should spectral measures be taken in the second pass?}

\item{...}{Further arguments passed to \link{run_praatsauce}.}
}
\value{
A data frame with columns giving information about the
sound file imported from the name of the sound file, information about
the time at which measures were estimated, and the estimated measures
themselves.
}
\description{
Shell wrapper calling PraatSauce to measure F0 in order to determine
suitable minimum and maximum limits, then reruns PraatSauce with supplied
arguments and the new F0 limits. Usually called from
\link{run_ps_dynamic_minmax}.
}
\examples{
#not now
}
