% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuhelpeR-package.R
\docType{package}
\name{emuhelpeR-package}
\alias{emuhelpeR}
\alias{emuhelpeR-package}
\title{emuhelpeR: Convenience functions for working with emuR}
\description{
Various convenience functions for working with the phonetic corpus tools in the emuR package.
}
\author{
\strong{Maintainer}: Rasmus Puggaard-Rode \email{r.puggaard@phonetik.uni-muenchen.de} (\href{https://orcid.org/0000-0003-4522-9987}{ORCID})

}
\keyword{internal}
