% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f0_proc.R
\name{f0_proc}
\alias{f0_proc}
\title{Processing fundamental frequency and dependencies}
\usage{
f0_proc(
  df,
  f0col = "F0",
  dep = NULL,
  speaker = NULL,
  group_var = NULL,
  timing_rm = NULL,
  report = TRUE
)
}
\arguments{
\item{df}{A data frame containing a column with fundamental frequency
measurements and columns with any other information to be accounted for.}

\item{f0col}{A string containing the name of the column in \code{df} that contains
fundamental frequency values. Default is \code{F0}.}

\item{dep}{One or more strings containing the name of columns in \code{df}
containing measures that are directly dependent on fundamental frequency.
Optional; default is \code{NULL}.}

\item{speaker}{An optional string giving the name of the column in \code{df}
containing speakers id's. Default is \code{NULL}, in which case z-score
normalization will be done on the basis of the data at large and no rescaled
values will be returned.}

\item{group_var}{One or more strings containing the names of columns in \code{df}
to be used as grouping variables for automatic outlier removal.
Optional; default is \code{NULL}, in which case automatic outlier removal will
be based on means and standard deviation in the data at large.}

\item{timing_rm}{An optional list with two arguments:
\itemize{
\item A string containing the label associated with a boundary in the data.
Values that are measured sufficiently far from this boundary are recoded as
\code{NA}. \code{f0_proc} assumes that this label is stored in the column \code{labels} in
a data frame generated with \code{\link[emuR:get_trackdata]{emuR::get_trackdata()}}.
\item A number indicating the distance from this boundary at which measurements
should be ignored. See example below.
}

Optional; default is \code{NULL}.}

\item{report}{A Boolean. If \code{TRUE} (default), prints a message stating
how many NAs were in the track originally and how many values were
recoded as \code{NA} during automated outlier removal. If there are F0 dependencies,
also prints how many values were recoded as \code{NA} in those.}
}
\value{
A data frame identical to \code{df} with outliers coded as \code{NA} and
with the added columns \code{zF0} and \code{normF0}, and corresponding column for
values that depend on F0. The column with F0 values are renamed \code{F0}.
}
\description{
\code{f0_proc} performs automatic outlier removal of fundamental frequency
values and dependent values, normalizes the resulting values, and rescales to
the original frequency scale.
}
\details{
Zero values are recoded as \code{NA}. Values are
recoded as \code{NA} if they fall outside of three standard deviations of the mean
value within the same group. If the data contain other measures that are
dependent on F0, e.g. spectral measures, then the corresponding values
of these are also coded as \code{NA}.

\code{f0_proc} also calculates by-speaker z-score normalization and stores it
in a separate column, then rescales the z-scores based on the overall mean
and SD values.

Optionally, values that are measured too far away from a boundary specified
by the user are coded as \code{NA}. This option assumes that the data frame
was generated using the \code{\link[emuR:get_trackdata]{emuR::get_trackdata()}} function to import data
from an EMU-SDMS database.
}
\examples{
colnames(ssff_data)
head(ssff_data$praatF0, 20)
head(ssff_data$H1H2c, 20)
sum(is.na(ssff_data$praatF0))
sum(is.na(ssff_data$H1H2c))
#Uses both speaker and vowel variables for outlier removal; F0 values and
#dependents are recoded as NA if they are more than 250 ms from boundaries
#labeled 'cl'.
x <- f0_proc(df=ssff_data, f0col='praatF0', dep='H1H2c', speaker='speaker',
group_var=c('speaker', 'vowel'), timing_rm=list('cl', 250))
colnames(x)
head(x$F0, 20)
head(x$zF0, 20)
head(x$normF0, 20)
head(x$zH1H2c, 20)
head(x$normH1H2c, 20)
sum(is.na(x$F0))
sum(is.na(x$H1H2c))
}
\seealso{
Some of the functionality in \code{f0_proc} assumes that the user has
generated the raw data using \link{emuR} and EMU-SDMS.

\code{f0_proc} is incorporated in the function \code{\link[=import_ssfftracks]{import_ssfftracks()}} which
automatically imports F0 values and other available measures from an EMU
database, performs outlier removal, normalizes values, and rescales.

More information about the example data can be found in \link{seg_list} and
\link{ssff_data}.
}
