% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_proc.R
\name{fn_proc}
\alias{fn_proc}
\title{Processing formants and dependencies}
\usage{
fn_proc(
  df,
  fncol = c("F1", "F2"),
  f0col = "F0",
  fndep = NULL,
  speaker = NULL,
  group_var = NULL,
  report = TRUE
)
}
\arguments{
\item{df}{#' @param df A data frame containing one or more columns with
formant measurements and columns with any other information to be accounted
for.}

\item{fncol}{One or more strings containing the names of columns in \code{df} that
contain formant values. Default is \code{c('F1', 'F2')}.}

\item{f0col}{A string containing the name of the column in \code{df} that contains
fundamental frequency values. Default is \code{F0}; this column will automatically
be renamed \code{F0} if the data has already been processed using \code{\link[=f0_proc]{f0_proc()}}.
Only used if formant dependencies are also being processed; see \code{fndep}.}

\item{fndep}{A list where each element contains two strings:
\itemize{
\item The name of a measure that is directly dependent on a specific formant.
\item Which formant the measure is dependent on.
}

The structure is as follows:
\verb{fndep=list(c('F1dep', 'F1'), c('F2dep', 'F2')}

Optional; default is \code{NULL}.}

\item{speaker}{An optional string giving the name of the column in \code{df}
containing speakers id's. Default is \code{NULL}, in which case z-score
normalization will be done on the basis of the data at large and no rescaled
values will be returned.}

\item{group_var}{One or more strings containing the names of columns in \code{df}
to be used as grouping variables for automatic outlier removal.
Optional; default is \code{NULL}, in which case automatic outlier removal will
be based on means and standard deviation in the data at large.}

\item{report}{A Boolean. If \code{TRUE} (default), prints a message stating
how many values in each track were
recoded as \code{NA} during automated outlier removal.}
}
\value{
A data frame identical to \code{df} with formant outliers coded as \code{NA} and
with the added columns \verb{zF\{n\}} and \verb{normF\{n\}}, and corresponding columns for
values that depend on formants. The columsn with formant values are renamed
\code{F1}, \code{F2}, etc.
}
\description{
\code{fn_proc} performs automatic outlier removal of formant
values and dependent values, normalizes the resulting values, and rescales to
the original frequency scale.
}
\details{
Zero values are recoded as \code{NA}. Values are
recoded as \code{NA} if they fall outside of three standard deviations of the mean
value within the same group. If the data contain other measures that are
dependent on specific formant values, e.g. spectral measures, then the
corresponding values of these are also coded as \code{NA}. This assumes that
these spectral measures are also dependent on F0, and that F0 has already
been processed using \code{\link[=f0_proc]{f0_proc()}}; this will also ensure that variables are
labeled correctly. Spectral measures which depend on e.g. F1 will be recoded
as \code{NA} if either the corresponding F0 measurement or the corresponding F1
measurement are labeled \code{NA}.

\code{fn_proc} also calculates by-speaker z-score normalization and stores it
in a separate column, then rescales the z-scores based on the overall mean
and SD values.

\code{fn_proc} is intended for formants, but values other than formants can
also be passed to the function if the user wishes to perform automated
outlier removal and/or normalization and rescaling on those.
}
\examples{
colnames(ssff_data)
head(ssff_data$praatF3, 20)
head(ssff_data$H1A3c, 20)
sum(is.na(ssff_data$praatF0))
sum(is.na(ssff_data$praatF3))
sum(is.na(ssff_data$H1A3c))
#see ?f0_proc
y <- f0_proc(df=ssff_data, f0col='praatF0', dep='H1H2c', speaker='speaker',
group_var=c('speaker', 'vowel'), timing_rm=list('cl', 250))
x <- fn_proc(df=y, fncol=c('praatF1', 'praatF2', 'praatF3'),
fndep=list(c('H1A1c', 'F1'), c('H1A3c', 'F3')),
speaker='speaker', group_var=c('speaker', 'vowel'))
head(x$F3, 20)
head(x$H1A3c, 20)
head(x$zF3, 20)
head(x$normF3, 20)
sum(is.na(x$F0))
sum(is.na(x$F3))
sum(is.na(x$H1A3c))
}
\seealso{
\code{fn_proc} should be functional regardless of how the raw data is
calculated, but the function is intended for raw data generated using \link{emuR}
and EMU-SDMS.

Some of the functionality in \code{fn_proc} assumes that the data has already
been processed using \code{\link[=f0_proc]{f0_proc()}}.

\code{fn_proc} is incorporated in the function \code{\link[=import_ssfftracks]{import_ssfftracks()}} which
automatically imports formant values and other available measures from an EMU
database, performs outlier removal, normalizes values, and rescales.

More information about the example data can be found in \link{seg_list} and
\link{ssff_data}.
}
