% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments2ssff.R
\name{moments2ssff}
\alias{moments2ssff}
\title{Add first four spectral moments as time series to EMU database}
\usage{
moments2ssff(db_handle, freq_ceiling = 8000, step = 5)
}
\arguments{
\item{db_handle}{The handle of an EMU database which is already loaded in R.}

\item{freq_ceiling}{Maximum frequency to be included in the spectra.
Default is \code{8000}, in which case the \code{snd_vector} is resampled to 16 kHz.}

\item{step}{Duration of time steps over which to generate spectra in ms.
Default is \code{5}}
}
\description{
Given a loaded EMU database, for each sound file will calculate the first
four spectral moments from DFT spectra over equidistant time steps and add
them as SSFF tracks to the database.
}
\examples{
\dontrun{
datapath <- system.file('extdata/ps', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
moments2ssff(raw)
#To have a look at an SSFF track in EMU-SDMS, run e.g.
add_signal_canvas(raw, 'cog')
emuR::serve(raw)
}
}
\seealso{
This function calls \code{\link[=moments_ts]{moments_ts()}} which can be used to calculate
spectral moments as time series for individual sound files.
This function partially wraps functionality from the \code{emuR} package, in
particular the \link[emuR:add_files]{emuR::add_files}, \link[emuR:AddListRemoveSsffTrackDefinition]{emuR::add_ssffTrackDefinition},
and \link[emuR:moments]{emuR::moments} functions.

\code{\link[=dct2ssff]{dct2ssff()}} is a sibling function for DCT coefficients.
}
