% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normz.R
\name{normz}
\alias{normz}
\title{Normalize by speaker and rescale}
\usage{
normz(df, var, speaker = NULL)
}
\arguments{
\item{df}{A data frame containing a column with the variable to be normalized
and rescaled, and optionally a column containing a unique id for each speaker.}

\item{var}{A string giving the name of the column in \code{df} containing the
continuous variable to be normalized.}

\item{speaker}{An optional string giving the name of the column in \code{df}
containing speakers id's. Default is \code{NULL}, in which case z-score
normalization will be done on the basis of the data at large and no rescaled
values will be returned.}
}
\value{
A data frame identical to \code{df} with the added columns \verb{z\{var\}} and
\verb{norm\{var\}}.
}
\description{
Performs by-speaker z-score normalization on a given variable and stores it
in a separate column, then rescales the z-scores based on the overall mean
and SD values in the data.
}
\examples{
colnames(ssff_data)
unique(ssff_data$speaker)
head(ssff_data$eggF0, 20)
x <- normz(df=ssff_data, var='eggF0', speaker='speaker')
colnames(x)
head(x$zeggF0, 20)
head(x$normeggF0, 20)
}
\seealso{
More information about the example data can be found in \link{seg_list} and
\link{ssff_data}.
}
