% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_rm.R
\name{outlier_rm}
\alias{outlier_rm}
\title{Automated removal of outliers by group}
\usage{
outlier_rm(df, var, group_var = NULL, report = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{var}{A string corresponding to a column name in \code{df} where outliers
should be removed.}

\item{group_var}{One or more strings containing the names of columns in \code{df}
to be used as grouping variables.
Optional; default is \code{NULL}, in which case automatic outlier removal will
be based on means and standard deviation in the data at large.}

\item{report}{A Boolean. If \code{TRUE} (default), prints a message stating
how many values were recoded as \code{NA}.}
}
\value{
A data frame identical to \code{df} with outliers recoded as \code{NA}.
}
\description{
Recodes values as \code{NA} if they fall outside of three standard deviations of
the mean value within the same grouping variable.
}
\examples{
sum(is.na(ssff_data$praatF0))
x <- outlier_rm(df=ssff_data, var='praatF0', group_var=c('speaker', 'vowel'))
sum(is.na(x$praatF0))
}
\seealso{
More information about the example data can be found in \link{seg_list} and
\link{ssff_data}.
}
