% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatsauce2ssff.R
\name{praatsauce2ssff}
\alias{praatsauce2ssff}
\title{Convert PraatSauce output to SSFF tracks and add to EMU database}
\usage{
praatsauce2ssff(ps_output, db_handle, session_col = NULL)
}
\arguments{
\item{ps_output}{Data frame with named columns containing PraatSauce output.}

\item{db_handle}{The handle of an EMU database which is already loaded in R.}

\item{session_col}{String containing the name of a column in \code{ps_output}
which contains the names of emuDB sessions. Optional; default is NULL,
in case all sessions are named \verb{0000_ses}, which is the default name of
an EMU-SDMS session in case there is only one.}
}
\description{
All measurement tracks in PraatSauce output file are converted into SSFF
tracks and added to an existing EMU database.
}
\examples{
datapath <- system.file('extdata/ps', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
dplyr::glimpse(ps)
# To have a look at an SSFF track in EMU-SDMS, run
#sco <- emuR::get_signalCanvasesOrder(raw, 'default')
#emuR::set_signalCanvasesOrder(raw, 'default', c(sco, 'f0'))
#emuR::serve(raw)
}
\seealso{
This function partially wraps functionality from the \code{emuR} package, in
particular the \link[emuR:add_files]{emuR::add_files} and \link[emuR:AddListRemoveSsffTrackDefinition]{emuR::add_ssffTrackDefinition}
functions.

The function is built to process output from the PraatSauce suite of Praat
scripts, which can be downloaded here:
\url{https://github.com/kirbyj/praatsauce}

Since PraatSauce aims to be as similar as possible to the VoiceSauce suite of
MATLAB scripts, the function should also work for VoiceSauce output,
although this has not been tested.
}
