% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\docType{data}
\name{ps}
\alias{ps}
\title{Example PraatSauce output}
\format{
A data frame with 3546 rows and 14 columns containing sample PraatSauce
output from a small subset of segments recorded for Kirby et al. (2023).
}
\source{
Kirby, James, Marc Brunelle & Pittayawat Pittayaporn (2023)
Transphonologization of onset voicing: Revisiting Northern and Eastern
Kmhmu. Phonetica. DOI:\url{https://doi.org/10.1515/phon-2022-0029}.

\url{https://doi.org/10.17605/OSF.IO/WV6QZ}
}
\usage{
ps
}
\description{
Example PraatSauce output
}
\keyword{datasets}
