% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_signal_canvas.R
\name{reset_signal_canvas}
\alias{reset_signal_canvas}
\title{Reset signal canvas order in EMU database}
\usage{
reset_signal_canvas(db_handle)
}
\arguments{
\item{db_handle}{The handle of an EMU database which is already loaded in R.}
}
\description{
Sets the default signal canvas order in an EMU database to the standard
setting of waveform and spectrogram.
}
\examples{
datapath <- system.file('extdata/db', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
emuR::get_signalCanvasesOrder(raw, 'default')
add_signal_canvas(db_handle=raw, add='eggF0')
emuR::get_signalCanvasesOrder(raw, 'default')
reset_signal_canvas(db_handle=raw)
emuR::get_signalCanvasesOrder(raw, 'default')
}
\seealso{
Wraps the function \link[emuR:SetGetSignalCanvasesOrder]{emuR::set_signalCanvasesOrder}.
}
