% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ps_dynamic_minmax.R
\name{run_ps_dynamic_minmax}
\alias{run_ps_dynamic_minmax}
\title{Run PraatSauce with dynamically estimated pitch floor and ceiling}
\usage{
run_ps_dynamic_minmax(
  directory,
  formantMeasures = TRUE,
  spectralMeasures = TRUE,
  ...
)
}
\arguments{
\item{directory}{Root directory containing subdirectories for individual
speakers to be measured. Must be the full name, there's no
path extension.}

\item{formantMeasures}{Boolean; should formants be measured in the second
pass?}

\item{spectralMeasures}{Boolean; should other spectral mesures be calculated
in the second pass?}

\item{...}{Further arguments passed to \link{run_praatsauce}.}
}
\value{
A data frame with columns giving information about the
sound file imported from the name of the sound file, information about
the time at which measures were estimated, and the estimated measures
themselves.
}
\description{
Shell wrapper to run PraatSauce for multiple speakers with dynamically
estimated F0 limits, following De Looze and Hirst (2021).
The function will first call PraatSauce to estimate F0 only with pitch
floor at 60 Hz and pitch ceiling at 700 Hz. The first and third quantiles
are calculated for each speaker, and PraatSauce is rerun with supplied
arguments and pitch floor estimated as 3/4 of the first quantile of the
first pass, and pitch ceiling as 1½ of the third quantile of the first pass.
}
\examples{
#Not now
}
\seealso{
As mentioned above, this function is a wrapper for a suite of
Praat scripts. More information about the scripts can be found
on GitHub: \url{https://github.com/kirbyj/praatsauce}. Information about the
VoiceSauce scripts which PraatSauce is based on can be found here:
\url{http://www.phonetics.ucla.edu/voicesauce/}.

The function \code{\link[=bulk_run_praatsauce]{bulk_run_praatsauce()}} will run PraatSauce in bulk
for multiple speakers with arguments supplied by the user.

The function \code{\link[=praatsauce2ssff]{praatsauce2ssff()}} will convert measures from PraatSauce to
SSFF files and add them to EMU database.
}
