% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_list.R
\docType{data}
\name{seg_list}
\alias{seg_list}
\title{Example segment list}
\format{
A tibble with 10 rows and 28 columns containing information about a small
subset of segments recorded for Kirby et al. (2023). To explore the
data with EMU-SDMS, call the following:

\code{datapath <- system.file('extdata/db', package='emuhelpeR')}

\code{raw <- emuR::load_emuDB(datapath)}

\code{emuR::serve(raw)}
}
\source{
Kirby, James, Marc Brunelle & Pittayawat Pittayaporn (2023)
Transphonologization of onset voicing: Revisiting Northern and Eastern
Kmhmu. Phonetica. DOI:\url{https://doi.org/10.1515/phon-2022-0029}.

\url{https://doi.org/10.17605/OSF.IO/WV6QZ}
}
\usage{
seg_list
}
\description{
Example segment list
}
\keyword{datasets}
