% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_on_spec.R
\name{signal_on_spec}
\alias{signal_on_spec}
\title{Display SSFF track on top of spectrogram when viewing EMU database}
\usage{
signal_on_spec(db, track, perspective = "default")
}
\arguments{
\item{db}{The handle of an EMU database which is loaded into R.}

\item{track}{A string giving the name of the SSFF track name to display on
top of the spectrogram. See \code{emuR::list_ssffTrackDefinitions()}.}

\item{perspective}{A string giving the name of the perspective for which the
configuration should be changed. Default is \code{default}.}
}
\description{
Change the configuration of an EMU database perspective to display signal
from an SSFF track on top of the spectrogram.
}
\examples{
\dontrun{
datapath <- system.file('extdata/db', package='emuhelpeR')
raw <- emuR::load_emuDB(datapath)
emuR::list_ssffTrackDefinitions(raw)
signal_on_spec(db, track='praatF0')
emuR::serve(raw)
}
}
\seealso{
\code{clear_spec()} resets the perspective to no longer display an SSFF
track on top of the spectrogram.
}
