% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_db}
\alias{load_db}
\title{Load EMU-SDMS database object}
\usage{
load_db(recursive = TRUE)
}
\arguments{
\item{recursive}{Logical; default is \code{TRUE}. Whether or not to recursively
search subdirectories of \code{dir} for
EMU databases.}
}
\value{
A list containing an EMU-SDMS database object.
}
\description{
Convenience function for loading an EMU-SDMS database without specifying
its name whenever there is only one EMU-SDMS database in the working
directory and its subdirectories.
}
\examples{
# don't run
# x <- load_db()
# serve()
}
\seealso{
This is a wrapper for the function \code{\link[emuR:load_emuDB]{emuR::load_emuDB()}}.
}
