% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_db}
\alias{load_db}
\title{Load and serve EMU-SDMS database object}
\usage{
load_db(recursive = TRUE, serve = TRUE)
}
\arguments{
\item{recursive}{Logical; default is \code{TRUE}. Whether or not to recursively
search subdirectories of \code{dir} for
EMU databases.}

\item{serve}{Logical; whether or not to serve the database in a browser window.}
}
\value{
A list containing an EMU-SDMS database object.
}
\description{
Convenience function for loading and serving EMU-SDMS database in a single
command without specifying
its name whenever there is only one EMU-SDMS database in the working
directory and its subdirectories.
}
\examples{
\dontrun{
x <- load_db()
}
}
\seealso{
This is a wrapper for the functions \code{\link[emuR:load_emuDB]{emuR::load_emuDB()}} and
\code{\link[emuR:serve]{emuR::serve()}}.
}
