% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssff_data.R
\docType{data}
\name{ssff_data}
\alias{ssff_data}
\title{Example SSFF track data}
\format{
A tibble with 2823 rows and 43 columns containing information about a small
subset of segments recorded for Kirby et al. (forthc.). To explore the
data with EMU-SDMS, call the following:

\code{datapath <- system.file('extdata/db', package='emuhelpeR')}

\code{raw <- emuR::load_emuDB(datapath)}

\code{emuR::serve(raw)}
}
\source{
Kirby, James, Marc Brunelle & Pittayawat Pittayaporn (forthc.)
Transphonologization of onset voicing: Revisiting Northern and Eastern
Kmhmu. To be published in Phonetica.

\url{https://doi.org/10.17605/OSF.IO/WV6QZ}
}
\usage{
ssff_data
}
\description{
Example SSFF track data
}
\keyword{datasets}
